+++ SpringDemoDB/src/main/java/com/SpringDemoDB/SpringDemoDB/Services/BookServiceImplementation.java
@@ -38,7 +38,7 @@ public class BookServiceImplementation implements BookService {
     }
 
     @Override
-    public void deleteBook (Long id) {
+    public void deleteBook (int id) {
         bookrepository.deleteById(id);
     }
